fix: route claimed Option input and terminal links coherently
