+++ CLI/cmux_open.swift
@@ -1470,19 +1470,7 @@ extension CMUXCLI {
            scheme == "http" || scheme == "https" {
             return .url(url.absoluteString, defaultFocus: true)
         }
-
-        let resolved: String
-        if let reference = TerminalPathResolver().resolveOpenURLFileReference(
-            raw,
-            cwd: FileManager.default.currentDirectoryPath
-        ) {
-            // The socket file.open contract is path-only. Normalize the
-            // location suffix here so `cmux open path:line` opens the file;
-            // terminal Cmd-click retains the location for the editor path.
-            resolved = reference.path
-        } else {
-            resolved = resolvePath(raw)
-        }
+        let resolved = TerminalPathResolver().resolveOpenURLFileReference(raw, cwd: FileManager.default.currentDirectoryPath)?.path ?? resolvePath(raw)
         var isDir: ObjCBool = false
         guard FileManager.default.fileExists(atPath: resolved, isDirectory: &isDir) else {
             throw CLIError(message: "Path does not exist: \(resolved)")
