+++ Packages/macOS/CmuxTerminalCore/Tests/CmuxTerminalCoreTests/TerminalPointerGestureStateTests.swift
@@ -5,24 +5,28 @@ import Testing
     @Test func completionUsesPressModifiers() throws {
         var state = TerminalPointerGestureState()
         state.begin(windowNumber: 7, timestamp: 10, modifierFlagsRawValue: 0x0010_0000, permitsLinkActivation: true)
-        let completion = try #require(state.complete(windowNumber: 7, timestamp: 11))
+        let rawCompletion = state.complete(windowNumber: 7, timestamp: 11)
+        let completion = try #require(rawCompletion)
         #expect(completion.modifierFlagsRawValue == 0x0010_0000)
         #expect(completion.permitsLinkActivation)
         #expect(!state.hasPendingRelease)
-        #expect(state.complete(windowNumber: 7, timestamp: 12) == nil)
+        let repeatedCompletion = state.complete(windowNumber: 7, timestamp: 12)
+        #expect(repeatedCompletion == nil)
     }
 
     @Test func releaseFromAnotherWindowClearsAuthorization() {
         var state = TerminalPointerGestureState()
         state.begin(windowNumber: 7, timestamp: 10, modifierFlagsRawValue: 0, permitsLinkActivation: true)
-        #expect(state.complete(windowNumber: 8, timestamp: 11) == nil)
+        let completion = state.complete(windowNumber: 8, timestamp: 11)
+        #expect(completion == nil)
         #expect(!state.hasPendingRelease)
     }
 
     @Test func staleReleaseClearsAuthorization() {
         var state = TerminalPointerGestureState()
         state.begin(windowNumber: 7, timestamp: 10, modifierFlagsRawValue: 0, permitsLinkActivation: true)
-        #expect(state.complete(windowNumber: 7, timestamp: 9) == nil)
+        let completion = state.complete(windowNumber: 7, timestamp: 9)
+        #expect(completion == nil)
         #expect(!state.hasPendingRelease)
     }
 
@@ -30,7 +34,8 @@ import Testing
         var state = TerminalPointerGestureState()
         state.begin(windowNumber: 7, timestamp: 10, modifierFlagsRawValue: 0, permitsLinkActivation: true)
         state.cancel()
-        #expect(state.complete(windowNumber: 7, timestamp: 20) == nil)
+        let completion = state.complete(windowNumber: 7, timestamp: 20)
+        #expect(completion == nil)
     }
 
     @Test func pointerExitRetainsReleaseButRevokesActivation() throws {
@@ -38,7 +43,8 @@ import Testing
         state.begin(windowNumber: 7, timestamp: 10, modifierFlagsRawValue: 0, permitsLinkActivation: true)
         state.invalidateLinkActivation()
         #expect(state.hasPendingRelease)
-        let completion = try #require(state.complete(windowNumber: 7, timestamp: 11))
+        let rawCompletion = state.complete(windowNumber: 7, timestamp: 11)
+        let completion = try #require(rawCompletion)
         #expect(!completion.permitsLinkActivation)
     }
 
@@ -46,7 +52,8 @@ import Testing
         var state = TerminalPointerGestureState()
         state.begin(windowNumber: 7, timestamp: 10, modifierFlagsRawValue: 1, permitsLinkActivation: true)
         state.begin(windowNumber: 9, timestamp: 20, modifierFlagsRawValue: 2, permitsLinkActivation: false)
-        let completion = try #require(state.complete(windowNumber: 9, timestamp: 21))
+        let rawCompletion = state.complete(windowNumber: 9, timestamp: 21)
+        let completion = try #require(rawCompletion)
         #expect(completion.modifierFlagsRawValue == 2)
         #expect(!completion.permitsLinkActivation)
     }
