+++ Sources/GhosttyTerminalView.swift
@@ -22,122 +22,6 @@ import CMUXMobileCore
 import IOSurface
 import UniformTypeIdentifiers
 
-enum GhosttyStartupAppearancePreviewProfile: String, CaseIterable, Identifiable {
-    case realUserConfig
-    case freshInstall
-    case userThemePair
-    case userSingleTheme
-    case userExplicitColors
-
-    var id: String { rawValue }
-
-    var displayName: String {
-        switch self {
-        case .realUserConfig:
-            return String(
-                localized: "debug.startupAppearance.profile.realUserConfig.title",
-                defaultValue: "Real User Config"
-            )
-        case .freshInstall:
-            return String(
-                localized: "debug.startupAppearance.profile.freshInstall.title",
-                defaultValue: "Fresh Install"
-            )
-        case .userThemePair:
-            return String(
-                localized: "debug.startupAppearance.profile.userThemePair.title",
-                defaultValue: "User Light/Dark Theme"
-            )
-        case .userSingleTheme:
-            return String(
-                localized: "debug.startupAppearance.profile.userSingleTheme.title",
-                defaultValue: "User Single Theme"
-            )
-        case .userExplicitColors:
-            return String(
-                localized: "debug.startupAppearance.profile.userExplicitColors.title",
-                defaultValue: "User Explicit Colors"
-            )
-        }
-    }
-
-    var detail: String {
-        switch self {
-        case .realUserConfig:
-            return String(
-                localized: "debug.startupAppearance.profile.realUserConfig.detail",
-                defaultValue: "Loads your actual Ghostty and cmux config files."
-            )
-        case .freshInstall:
-            return String(
-                localized: "debug.startupAppearance.profile.freshInstall.detail",
-                defaultValue: "No user Ghostty settings, so cmux applies its managed default colors."
-            )
-        case .userThemePair:
-            return String(
-                localized: "debug.startupAppearance.profile.userThemePair.detail",
-                defaultValue: "Simulates a user with an explicit light/dark Ghostty theme."
-            )
-        case .userSingleTheme:
-            return String(
-                localized: "debug.startupAppearance.profile.userSingleTheme.detail",
-                defaultValue: "Simulates a user with one Ghostty theme applied in both appearances."
-            )
-        case .userExplicitColors:
-            return String(
-                localized: "debug.startupAppearance.profile.userExplicitColors.detail",
-                defaultValue: "Simulates a user with direct terminal color settings and no theme."
-            )
-        }
-    }
-
-    var loadsRealUserConfig: Bool {
-        self == .realUserConfig
-    }
-
-    func previewConfigContents(
-        preferredColorScheme: GhosttyConfig.ColorSchemePreference = GhosttyConfig.currentColorSchemePreference()
-    ) -> String? {
-        switch self {
-        case .realUserConfig:
-            return nil
-        case .freshInstall:
-            return GhosttyConfig.cmuxDefaultThemeConfigContents(
-                preferredColorScheme: preferredColorScheme
-            )
-        case .userThemePair:
-            return "theme = light:Catppuccin Latte,dark:Catppuccin Mocha"
-        case .userSingleTheme:
-            return "theme = Catppuccin Mocha"
-        case .userExplicitColors:
-            return """
-            background = #101820
-            foreground = #F4F7F7
-            cursor-color = #FEE715
-            cursor-text = #101820
-            selection-background = #28536B
-            selection-foreground = #F4F7F7
-            palette = 0=#101820
-            palette = 1=#C14953
-            palette = 2=#47A025
-            palette = 3=#D9A441
-            palette = 4=#2E86AB
-            palette = 5=#9B5DE5
-            palette = 6=#00A6A6
-            palette = 7=#D6D6D6
-            palette = 8=#5C6672
-            palette = 9=#FF6B6B
-            palette = 10=#7BD88F
-            palette = 11=#FFD166
-            palette = 12=#54C6EB
-            palette = 13=#C77DFF
-            palette = 14=#4ECDC4
-            palette = 15=#FFFFFF
-            """
-        }
-    }
-}
-
 enum GhosttyStartupAppearancePreviewState {
     #if DEBUG
     // The selected debug preview profile. Backed by the CmuxTerminalCore seam
@@ -3610,6 +3494,7 @@ class GhosttyApp {
             }
         case GHOSTTY_ACTION_OPEN_URL:
             let openUrl = action.action.open_url
+            let isTerminalLink = openUrl.kind == GHOSTTY_ACTION_OPEN_URL_KIND_UNKNOWN
             guard let cstr = openUrl.url else { return false }
             let urlString = String(
                 data: Data(bytes: cstr, count: Int(openUrl.len)),
@@ -3622,6 +3507,12 @@ class GhosttyApp {
                 workingDirectory: surfaceView.currentDirectoryActionDispatcher.directorySnapshot()
             )
             return performOnMain {
+                // Link callbacks must belong to one intentional pointer
+                // release. Text/HTML exports carry their own explicit action
+                // kind and can also be opened by a configured key binding.
+                if isTerminalLink, !surfaceView.consumeTerminalLinkOpenAuthorization() {
+                    return true
+                }
                 surfaceView.recordCommandClickReleaseRuntimeOutcome(.openURL)
                 return TerminalLinkOpenCoordinator().open(request)
             }
@@ -3857,6 +3748,8 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
     private let commandClickReleaseRouter = TerminalCommandClickReleaseRouter()
     private var commandClickReleaseRoutingActive = false
     private var commandClickReleaseRuntimeOutcome: TerminalCommandClickReleaseRouter.RuntimeOutcome?
+    private var commandClickReleaseCanOpenURL = false
+    private var terminalPointerGesture = TerminalPointerGestureState()
     private var ghosttyMouseShape: ghostty_action_mouse_shape_e = GHOSTTY_MOUSE_SHAPE_TEXT
     private static func ghosttyMouseCursor(for shape: ghostty_action_mouse_shape_e) -> NSCursor {
         switch shape {
@@ -4216,6 +4109,7 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
     fileprivate var isVisibleInUI: Bool { visibleInUI }
     fileprivate func setVisibleInUI(_ visible: Bool) {
         visibleInUI = visible
+        if !visible { terminalPointerGesture.cancel() }
     }
 
     override init(frame frameRect: NSRect) {
@@ -4591,6 +4485,7 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
         _ session: GhosttyMouseSessionLedger.Session
     ) -> Bool {
         let finished = ghosttyMouseSessionLedger.finish(session)
+        if finished, session.button == .left { terminalPointerGesture.cancel() }
         removeMouseUpEventMonitorIfUnused()
         if session.button == .right {
             removeContextMenuEndObserver()
@@ -4599,6 +4494,8 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
     }
 
     private func resetGhosttyMouseButtonTracking() {
+        terminalPointerGesture.cancel()
+        commandClickReleaseCanOpenURL = false
         ghosttyMouseSessionLedger.invalidate()
         removeMouseUpEventMonitorIfUnused()
         removeContextMenuEndObserver()
@@ -6807,10 +6704,11 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
                 keyCode: event.keyCode
             ) ?? event
         }
-        let textInputEvent = textInputInterpretationEvent(
-            original: event,
-            translated: translationEvent
-        )
+        // Ghostty's translation modifiers are the source of truth for both
+        // terminal encoding and AppKit text interpretation. Showing AppKit a
+        // second, Option-bearing event here reintroduces dead-key composition
+        // for keys that `macos-option-as-alt` intentionally claims.
+        let textInputEvent = translationEvent
 
         // Set up text accumulator for interpretKeyEvents
         keyTextAccumulator = []
@@ -7544,6 +7442,8 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
     }
 
     fileprivate func deferReleaseAllGhosttyMouseButtons(reason: String) {
+        terminalPointerGesture.cancel()
+        commandClickReleaseCanOpenURL = false
         guard currentGhosttyMouseSurfaceIdentity != nil,
               !ghosttyMouseSessionLedger.activeButtons.isEmpty else {
             resetGhosttyMouseButtonTracking()
@@ -7759,6 +7659,7 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
 
     override func mouseDown(with event: NSEvent) {
         if routeInputDuringClipboardRead(event) { return }
+        terminalPointerGesture.cancel()
         reconcileGhosttyMouseButtons(
             reason: "mouseDown.preflight",
             forceButtons: Set([.left])
@@ -7777,6 +7678,13 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
         cmuxDebugLog("terminal.mouseDown surface=\(terminalSurface?.id.uuidString.prefix(5) ?? "nil") mods=[\(debugModifierString(event.modifierFlags))] clickCount=\(event.clickCount) point=(\(String(format: "%.0f", debugPoint.x)),\(String(format: "%.0f", debugPoint.y)))")
         #endif
         let eventPoint = mouseState.localPoint
+        let pressFlags = event.modifierFlags.intersection(.deviceIndependentFlagsMask)
+        terminalPointerGesture.begin(
+            windowNumber: event.windowNumber,
+            timestamp: event.timestamp,
+            modifierFlagsRawValue: pressFlags.rawValue,
+            permitsLinkActivation: pressFlags.contains(.command) && bounds.contains(eventPoint)
+        )
         trackMousePointIfUsable(eventPoint)
         // Only update mouse position on the first click to prevent unwanted cursor
         // movement during double-click selection (issue #1698)
@@ -7802,6 +7710,7 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
     @discardableResult
     func forwardPendingLeftMouseDrag(with event: NSEvent) -> Bool {
         if routeInputDuringClipboardRead(event) { return true }
+        terminalPointerGesture.invalidateLinkActivation()
         synchronizeGhosttyMouseSurfaceIdentity()
         guard let surface,
               ghosttyMouseSessionLedger.hasSession(
@@ -7833,19 +7742,36 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
             return false
         }
         let point = mouseState.localPoint
+        let completion = terminalPointerGesture.complete(
+            windowNumber: event.windowNumber,
+            timestamp: event.timestamp
+        )
+        let releaseFlags = completion.map {
+            NSEvent.ModifierFlags(rawValue: $0.modifierFlagsRawValue)
+        } ?? []
+        let linkActivationAuthorized = completion?.permitsLinkActivation == true
+            && event.modifierFlags.contains(.command) && bounds.contains(point) && desiredFocus
         _ = dispatchCommandClickRelease(
             surface: surface,
             at: point,
-            modifierFlags: event.modifierFlags,
-            mouseMods: mouseState.mods
+            modifierFlags: releaseFlags,
+            mouseMods: mouseState.mods,
+            linkActivationAuthorized: linkActivationAuthorized
         )
         _ = finishGhosttyMouseSession(pendingSession)
         return true
     }
 
-    private func beginCommandClickReleaseRouting() {
+    private func beginCommandClickReleaseRouting(linkActivationAuthorized: Bool) {
         commandClickReleaseRoutingActive = true
         commandClickReleaseRuntimeOutcome = nil
+        commandClickReleaseCanOpenURL = linkActivationAuthorized
+    }
+
+    fileprivate func consumeTerminalLinkOpenAuthorization() -> Bool {
+        guard commandClickReleaseRoutingActive, commandClickReleaseCanOpenURL else { return false }
+        commandClickReleaseCanOpenURL = false
+        return true
     }
 
     /// Records a URL action only while its originating mouse release is active.
@@ -7863,6 +7789,7 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
         defer {
             commandClickReleaseRoutingActive = false
             commandClickReleaseRuntimeOutcome = nil
+            commandClickReleaseCanOpenURL = false
         }
         guard commandClickReleaseRoutingActive else {
             return ghosttyConsumed ? .consumed : .unhandled
@@ -7875,13 +7802,14 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
         surface: ghostty_surface_t,
         at point: NSPoint,
         modifierFlags: NSEvent.ModifierFlags,
-        mouseMods: ghostty_input_mods_e
+        mouseMods: ghostty_input_mods_e,
+        linkActivationAuthorized: Bool
     ) -> (
         consumed: Bool,
         runtimeOutcome: TerminalCommandClickReleaseRouter.RuntimeOutcome,
         resolution: WordPathResolution?
     ) {
-        beginCommandClickReleaseRouting()
+        beginCommandClickReleaseRouting(linkActivationAuthorized: linkActivationAuthorized)
         let consumed = sendGhosttyMouseButton(
             surface,
             state: GHOSTTY_MOUSE_RELEASE,
@@ -7889,11 +7817,11 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
             mods: mouseMods
         )
         let runtimeOutcome = finishCommandClickReleaseRouting(ghosttyConsumed: consumed)
-        let resolution = handleCommandClickRelease(
+        let resolution = linkActivationAuthorized ? handleCommandClickRelease(
             at: point,
             modifierFlags: modifierFlags,
             runtimeOutcome: runtimeOutcome
-        )
+        ) : nil
         return (consumed, runtimeOutcome, resolution)
     }
 
@@ -8539,7 +8467,8 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
             surface: surface,
             at: clampedPoint,
             modifierFlags: flags,
-            mouseMods: mods
+            mouseMods: mods,
+            linkActivationAuthorized: true
         )
 
         var payload: [String: Any] = [
@@ -8588,7 +8517,8 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
             surface: surface,
             at: clampedPoint,
             modifierFlags: flags,
-            mouseMods: commandMods
+            mouseMods: commandMods,
+            linkActivationAuthorized: true
         )
         flagsChanged(with: cmdUp)
 
@@ -8979,6 +8909,7 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
     }
 
     override func mouseExited(with event: NSEvent) {
+        terminalPointerGesture.invalidateLinkActivation()
         if routeInputDuringClipboardRead(event) { return }
         reconcileGhosttyMouseButtons(reason: "mouseExited")
         if wordPathHoverActive {
